feat: 메모리 캐시 시스템 구현으로 API 성능 최적화
- 메모리 기반 캐시 유틸리티 추가 (backend/app/utils/cache.py)
- 모든 읽기 엔드포인트에 캐시 적용 (ETF 목록, 가격, 뉴스, 거래흐름 등)
- 데이터 수집 후 자동 캐시 무효화 로직 구현
- 캐시 통계 조회 및 수동 삭제 API 추가
- 프론트엔드 staleTime 30초로 조정하여 실시간성 향상

+++ backend/app/config.py
@@ -34,8 +34,8 @@ class Config:
     # Scheduler Settings
     SCRAPING_INTERVAL_MINUTES = int(os.getenv("SCRAPING_INTERVAL_MINUTES", "3"))
 
-    # Cache Settings (현재 미구현, 향후 Redis 또는 메모리 캐시 도입 시 사용)
-    CACHE_TTL_MINUTES = int(os.getenv("CACHE_TTL_MINUTES", "3"))
+    # Cache Settings (메모리 캐시 구현됨)
+    CACHE_TTL_MINUTES = float(os.getenv("CACHE_TTL_MINUTES", "3"))
     
     # Stock Configuration File Path
     STOCK_CONFIG_PATH = os.getenv(
